+++ src/Domain/Storage/Filesystem/FilePointer.php
@@ -21,13 +21,8 @@ class FilePointer implements EventProvider
 
     /**
      * @ORM\Id
-     * @ORM\Column(type="integer", name="id")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    private ?int $internalId = null;
-
-    /**
-     * @ORM\Column(type="string", unique=true, length=36, nullable=false, name="uuid")
+     * @ORM\Column(name="uuid", type="string", length=36)
+     * @ORM\GeneratedValue(strategy="NONE")
      */
     private string $id;