+++ src/Infrastructure/Domain/File/Doctrine/DoctrineFileRepository.php
@@ -4,12 +4,12 @@ namespace Becklyn\Ddd\FileStore\Infrastructure\Domain\File\Doctrine;
 
 use Becklyn\Ddd\Events\Domain\EventCreatorCapabilities;
 use Becklyn\Ddd\Events\Domain\EventRegistry;
-use Becklyn\Ddd\Identity\Domain\AggregateId;
 use Becklyn\Ddd\FileStore\Domain\File\File;
 use Becklyn\Ddd\FileStore\Domain\File\FileDeleted;
 use Becklyn\Ddd\FileStore\Domain\File\FileId;
 use Becklyn\Ddd\FileStore\Domain\File\FileNotFoundException;
 use Becklyn\Ddd\FileStore\Domain\File\FileRepository;
+use Becklyn\Ddd\Identity\Domain\AggregateId;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ObjectRepository;